# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ArMenu.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/menus/{tenant}/events": {
      "post": {
        "tags": [
          "Public menu"
        ],
        "summary": "Counts guest events on the menu: views, opened dishes, loaded models, AR starts. No guest identifier is accepted or stored.",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Slug of the business, as printed in its QR code links.",
            "required": true,
            "schema": {
              "maxLength": 63,
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuEventsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manage/statistics": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Daily guest activity and the dishes guests open most, over the last days (7 by default, at most 90).",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuStatisticsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/v1/auth/sign-in": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Signs a user-name account in to its own business, or the administrator to the platform; says which.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserNameSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceAccessTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "The signed-in member and the tenant the access token is scoped to.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/v1/tenants/{tenant}/auth/sign-in": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Signs a member in and returns an access token; the refresh token is set as an HttpOnly cookie.",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Slug of the business, as printed in its QR code links.",
            "required": true,
            "schema": {
              "maxLength": 63,
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant}/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rotates the refresh token cookie and returns a new access token.",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Slug of the business, as printed in its QR code links.",
            "required": true,
            "schema": {
              "maxLength": 63,
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant}/auth/sign-out": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Ends the session and clears the refresh token cookie.",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Slug of the business, as printed in its QR code links.",
            "required": true,
            "schema": {
              "maxLength": 63,
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessTokenResponse": {
        "required": [
          "accessToken",
          "expiresAt",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          },
          "tokenType": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CurrentTenantResponse": {
        "required": [
          "id",
          "slug",
          "name",
          "defaultCulture",
          "supportedCultures",
          "currency",
          "timeZone",
          "logoPath",
          "logoUrl",
          "accentColor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "defaultCulture": {
            "type": "string"
          },
          "supportedCultures": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "currency": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "logoPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "logoUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "accentColor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CurrentUserResponse": {
        "required": [
          "id",
          "email",
          "emailVerified",
          "fullName",
          "role",
          "userName",
          "isPlatformAdmin",
          "tenant"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "userName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPlatformAdmin": {
            "type": "boolean"
          },
          "tenant": {
            "$ref": "#/components/schemas/CurrentTenantResponse"
          }
        },
        "description": "The signed-in person. `UserName` is how a user-name account signs in, and null for e-mail accounts, whose\n`Email` is then a real address; `IsPlatformAdmin` says whether the platform pages are open to them."
      },
      "DailyMenuStatistics": {
        "required": [
          "day",
          "menuViews",
          "dishOpens",
          "arStarts"
        ],
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "format": "date"
          },
          "menuViews": {
            "type": "integer",
            "format": "int64"
          },
          "dishOpens": {
            "type": "integer",
            "format": "int64"
          },
          "arStarts": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "DishStatistics": {
        "required": [
          "itemId",
          "name",
          "opens",
          "modelViews",
          "arStarts"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "opens": {
            "type": "integer",
            "format": "int64"
          },
          "modelViews": {
            "type": "integer",
            "format": "int64"
          },
          "arStarts": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "A dish's activity, most opened first. Deleted dishes are left out."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. `tenant.not_found`."
          },
          "traceId": {
            "type": "string",
            "description": "Correlates the response with server logs and traces."
          },
          "errorCodes": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Stable error codes per field (JSON path), for localized messages on the client."
          }
        }
      },
...
      "MenuEventInput": {
        "required": [
          "type",
          "itemId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/MenuEventType"
          },
          "itemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "MenuEventsRequest": {
        "required": [
          "events"
        ],
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuEventInput"
            }
          }
        }
      },
      "MenuEventType": {
        "enum": [
          "menu_viewed",
          "dish_opened",
          "model_viewed",
          "ar_started"
        ],
        "description": "What a guest did. Counted per day, never stored per guest."
      },
...
      "MenuStatisticsResponse": {
        "required": [
          "from",
          "to",
          "totals",
          "days",
          "dishes"
        ],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "totals": {
            "$ref": "#/components/schemas/MenuStatisticsTotals"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyMenuStatistics"
            }
          },
          "dishes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishStatistics"
            }
          }
        }
      },
      "MenuStatisticsTotals": {
        "required": [
          "menuViews",
          "dishOpens",
          "modelViews",
          "arStarts"
        ],
        "type": "object",
        "properties": {
          "menuViews": {
            "type": "integer",
            "format": "int64"
          },
          "dishOpens": {
            "type": "integer",
            "format": "int64"
          },
          "modelViews": {
            "type": "integer",
            "format": "int64"
          },
          "arStarts": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. `tenant.not_found`."
          },
          "traceId": {
            "type": "string",
            "description": "Correlates the response with server logs and traces."
          }
        }
      },
...
      "SignInRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "UserNameSignInRequest": {
        "required": [
          "userName",
          "password"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "WorkspaceAccessTokenResponse": {
        "required": [
          "accessToken",
          "expiresAt",
          "expiresIn",
          "workspace",
          "isPlatformAdmin"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          },
          "workspace": {
            "type": "string"
          },
          "isPlatformAdmin": {
            "type": "boolean"
          },
          "tokenType": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "A token response that also says where the session lives: `Workspace` is the slug whose auth endpoints\nrefresh and end it, and `IsPlatformAdmin` is true for the administrator, whose workspace is the platform."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token from sign-in or refresh. Tokens are scoped to one tenant.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}